Fail only if there are more errors than `maxAllowedErrors`
It is nice to have a clean lint but in reality for projects that take
jshint into use later in their life cleaning all the errors in one go
would be a huge task. It is easier to set a limit for the lint
warnings and steadily lower that limit.