extension is after _last_ period only
in file.expandMapping, fix destPath replace such that only the extension
is replaced, not anything after the first '.' of the last path segment.

In other words, if the source file is 'my/dir/path/test.single.js', only
the final '.js' should be considered the extension, not '.single.js',
and only that should be replaced.