Keep track of last run time and modify newer option
This adds support for a "newer" option for files.  If true, only files modified since the last run will be matched.  If false (the default), files are matched in the normal way.  If the newer option is "any", all files will be matched if any one of them has been modified since the last run.