Short is a reserved word
Short is a reserved word in [ECMAScript 2+](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Reserved_Words). As such, it can't be used an identifier, strictly. An easy solution is to put it in quotes. I've done that for the main grunt/cli.js file. It's not pretty, but it's valid.