Search ancestors for node_modules in grunt.loadNpmTasks
Fixes #1211.

Note: the reporter of the issue suggested using `require`. I initially tried
using that to implement this. However, I soon ran into issues, and I realized
that there is a fundamental issue: a module can only be `require`d if it
specifies a `main` file in its `package.json`. Many Grunt plugins do not do
this--so that approach is out.