Bump minimatch to 3.0.4 or later
Fix npm audit warning:
                   
  Moderate        ReDoS
  Package         brace-expansion
  Dependency of   grunt [dev]
  Path            grunt > minimatch > brace-expansion
  More info       https://nodesecurity.io/advisories/338 

It seems the ~ syntax is expanded by npm to newer versions automatically,
even on clean installs. However, it can be worked around by specifying
a newer version of minimatch directly in the project that uses grunt, because
the range does allow deduplication. Anyhow, let's fix this on the grunt side
instead.